The screen scale selection logic is reconstructed, and the WindowUtils class is used to centrally manage window-related functions
- Remove redundant code from AppDelegate and openterfaceApp and use the WindowUtils class to display the scale selector
- Update the notification mechanism for window sizes with new notification names to improve readability
- Optimize window update logic after user selects screen scale
